api_authorizer-requirements.txt
-i https://pypi.org/simple
aws-xray-sdk==2.4.2
future==0.18.1
jsonpickle==1.2
wrapt==1.11.2
google-auth==1.6.3
cachetools==3.1.1
pyasn1-modules==0.2.7
pyasn1==0.4.7
rsa==4.0
